+++ src/apis/o_auth_api.rs
@@ -172,7 +172,7 @@ pub async fn delete_o_auth_app(configuration: &configuration::Configuration, app
 }
 
 /// Get a page of OAuth 2.0 client applications authorized to access a user's account. ##### Permissions Must be authenticated as the user or have `edit_other_users` permission. 
-pub async fn get_authorized_o_auth_apps_for_user(configuration: &configuration::Configuration, user_id: &str, page: Option<i32>, per_page: Option<i32>) -> Result<Vec<crate::models::OAuthApp>, Error<GetAuthorizedOAuthAppsForUserError>> {
+pub async fn get_authorized_o_auth_apps_for_user(configuration: &configuration::Configuration, user_id: &str, page: Option<i64>, per_page: Option<i64>) -> Result<Vec<crate::models::OAuthApp>, Error<GetAuthorizedOAuthAppsForUserError>> {
     let local_var_configuration = configuration;
 
     let local_var_client = &local_var_configuration.client;
@@ -271,7 +271,7 @@ pub async fn get_o_auth_app_info(configuration: &configuration::Configuration, a
 }
 
 /// Get a page of OAuth 2.0 client applications registered with Mattermost. ##### Permissions With `manage_oauth` permission, the apps registered by the logged in user are returned. With `manage_system_wide_oauth` permission, all apps regardless of creator are returned. 
-pub async fn get_o_auth_apps(configuration: &configuration::Configuration, page: Option<i32>, per_page: Option<i32>) -> Result<Vec<crate::models::OAuthApp>, Error<GetOAuthAppsError>> {
+pub async fn get_o_auth_apps(configuration: &configuration::Configuration, page: Option<i64>, per_page: Option<i64>) -> Result<Vec<crate::models::OAuthApp>, Error<GetOAuthAppsError>> {
     let local_var_configuration = configuration;
 
     let local_var_client = &local_var_configuration.client;
