+++ src/models/channel_with_team_data.rs
@@ -38,9 +38,9 @@ pub struct ChannelWithTeamData {
     pub purpose: Option<String>,
     /// The time in milliseconds of the last post of a channel
     #[serde(rename = "last_post_at", skip_serializing_if = "Option::is_none")]
-    pub last_post_at: Option<i32>,
+    pub last_post_at: Option<i64>,
     #[serde(rename = "total_msg_count", skip_serializing_if = "Option::is_none")]
-    pub total_msg_count: Option<i32>,
+    pub total_msg_count: Option<i64>,
     /// Deprecated in Mattermost 5.0 release
     #[serde(rename = "extra_update_at", skip_serializing_if = "Option::is_none")]
     pub extra_update_at: Option<i64>,
@@ -54,7 +54,7 @@ pub struct ChannelWithTeamData {
     pub team_name: Option<String>,
     /// The time at which the team to which this channel belongs was last updated.
     #[serde(rename = "team_update_at", skip_serializing_if = "Option::is_none")]
-    pub team_update_at: Option<i32>,
+    pub team_update_at: Option<i64>,
     /// The data retention policy to which this team has been assigned. If no such policy exists, or the caller does not have the `sysconsole_read_compliance_data_retention` permission, this field will be null.
     #[serde(rename = "policy_id", skip_serializing_if = "Option::is_none")]
     pub policy_id: Option<String>,
