+++ src/models/config_analytics_settings.rs
@@ -14,7 +14,7 @@
 #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
 pub struct ConfigAnalyticsSettings {
     #[serde(rename = "MaxUsersForStatistics", skip_serializing_if = "Option::is_none")]
-    pub max_users_for_statistics: Option<i32>,
+    pub max_users_for_statistics: Option<i64>,
 }
 
 impl ConfigAnalyticsSettings {
